+++ client/src/main/java/com/microsoft/durabletask/OrchestrationRuntimeStatus.java
@@ -46,7 +46,12 @@ public enum OrchestrationRuntimeStatus {
     /**
      * The orchestration was scheduled but hasn't started running.
      */
-    PENDING;
+    PENDING,
+
+    /**
+     * The orchestration is in a suspended state.
+     */
+    SUSPENDED;
 
     static OrchestrationRuntimeStatus fromProtobuf(OrchestrationStatus status) {
         switch (status) {
@@ -64,6 +69,8 @@ public enum OrchestrationRuntimeStatus {
                 return TERMINATED;
             case ORCHESTRATION_STATUS_PENDING:
                 return PENDING;
+            case ORCHESTRATION_STATUS_SUSPENDED:
+                return SUSPENDED;
             default:
                 throw new IllegalArgumentException(String.format("Unknown status value: %s", status));
         }
@@ -85,6 +92,8 @@ public enum OrchestrationRuntimeStatus {
                 return ORCHESTRATION_STATUS_TERMINATED;
             case PENDING:
                 return ORCHESTRATION_STATUS_PENDING;
+            case SUSPENDED:
+                return ORCHESTRATION_STATUS_SUSPENDED;
             default:
                 throw new IllegalArgumentException(String.format("Unknown status value: %s", status));
         }